requirements.txt
python-telegram-bot
requests
psycopg2-binary
tweepy